python/requirements.txt
azure-core==1.34.0
azure-servicebus==7.14.2
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dotenv==0.9.9
idna==3.10
isodate==0.7.2
pycparser==2.22
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
